styleAI/requirements.txt
flask
flask-cors
python-dotenv
Pillow
google-generativeai
